version 0.9.43

+++ src/constrct.cpp
@@ -896,18 +896,18 @@ double construct::mdlAux(constructNode *Node)
          return mdlAux(Node->left) + mdlAux(Node->right) ;
 
       case cnCONTattribute: // summing and multiplying
-         return log2(gFT->noNumeric-1) ;
+         return log2((double)(gFT->noNumeric-1)) ;
       
       case cnCONTattrValue:
       {
         double intValue = gFT->valueInterval[Node->attrIdx]/gFT->opt->mdlErrorPrecision ;
         if (intValue < 1.0)
           intValue = 1.0 ;
-        return  log2(gFT->noAttr) + 2*log2(intValue) ;
+        return  log2((double)gFT->noAttr) + 2*log2((double)intValue) ;
       }
       case  cnDISCattrValue:
-        return log2(gFT->noAttr) +
-               log2(gFT->AttrDesc[gFT->DiscIdx[Node->attrIdx]].NoValues) ;
+        return log2((double)gFT->noAttr) +
+               log2((double)gFT->AttrDesc[gFT->DiscIdx[Node->attrIdx]].NoValues) ;
 
       case cnDISCattribute:
       default: 
@@ -927,11 +927,11 @@ double construct::mdlAux(constructNode *Node)
 // *********************************************************************
 double construct::mdlConstructCode() 
 {
-   double code = log2(no1bits(gFT->opt->constructionMode)) ;
+   double code = log2((double)no1bits(gFT->opt->constructionMode)) ;
    switch (compositionType)
    {
      case cSINGLEattribute:
-          code += log2(gFT->noAttr) ;
+          code += log2((double)gFT->noAttr) ;
           if (countType == aDISCRETE)
           {
 //#if defined(DEBUG)
@@ -958,7 +958,7 @@ double construct::mdlConstructCode()
      case cSUM:
      case cPRODUCT:
      case cXofN:
-         code += log2(degreesOfFreedom()) ;
+         code += log2((double)degreesOfFreedom()) ;
          code += mdlAux() ;       
          break ;
      default:
